+++ src/runtime/mod.rs
@@ -1,3 +1,5 @@
+mod docker;
 mod native;
 
+pub use docker::{OciApplicationSpec, OciError, OciSession};
 pub use native::{ApplicationSpec, NativeError, NativeSession, Xvfb};